refactor code and fix the issue related to the verification code that keeps equal when we make a request to get a new one
