+++ src/app.ts
@@ -1,3 +1,3 @@
-import Server from '../server';
+import Server from './server';
 const server = new Server();
 server.listen();
