+++ src/config/database.ts
@@ -1,6 +1,6 @@
 import { Sequelize } from 'sequelize';
 import 'colors';
-import { config } from '../config/config';
+import { config } from './config';
 
 export const sequelize = new Sequelize(config.DATABASE_NAME, config.DATABASE_USER, config.DATABASE_PASSWORD, {
 	dialect: 'mssql',
