+++ src/config/index.ts
@@ -2,3 +2,4 @@ export * from './CustomError';
 export * from './config';
 export * from './i18n';
 export * from './swagger';
+export * from './database';
