+++ src/middleware/validate-token.ts
