+++ src/middleware/validatorEnpoint.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from 'express';
 import { validationResult } from 'express-validator';
-import { StatusValue, StatusCode } from '../interface/example';
+import { StatusValue, StatusCode } from '../interface/auth.interface';
 
 const validateEnpoint = (req: Request, res: Response, next: NextFunction) => {
 	const error = validationResult(req);
@@ -8,7 +8,7 @@ const validateEnpoint = (req: Request, res: Response, next: NextFunction) => {
 		const data = error.mapped();
 		return res
 			.status(StatusCode.BadRequest)
-			.json({ status: StatusValue.Failed, data: {error: data} });
+			.json({ status: StatusValue.Failed, data: { error: data } });
 	}
 	next();
 };
