+++ src/repositories/index.ts
@@ -1,0 +1,5 @@
+export * from './model/rol.model';
+export * from './user/user.model';
+export * from './model/verification.model';
+export * as verificationRepository from './auth/verification.repository';
+export * as authRepository from './auth/auth.repository';
