+++ src/services/index.ts
@@ -1,0 +1,1 @@
+export * as authService from './auth/auth.service';
