--- src/utils/parse-messga-i18.ts
@@ -1,9 +1,0 @@
-export const parseMessageI18n = (message: any, req: any) => {
-	//validate message is object
-	if (typeof message === 'object') {
-		const { translationKey, translationParams } = message;
-		return  req.__(translationKey, translationParams);
-	}
-
-	return req.__(message);
-};
